+++ src/Publisher/Repository/StatsRepository.php
@@ -32,6 +32,7 @@ interface StatsRepository
     public const TYPE_VIEW = 'view';
     public const TYPE_VIEW_ALL = 'viewAll';
     public const TYPE_VIEW_INVALID_RATE = 'viewInvalidRate';
+    public const TYPE_VIEW_UNIQUE = 'viewUnique';
     public const TYPE_CLICK = 'click';
     public const TYPE_CLICK_ALL = 'clickAll';
     public const TYPE_CLICK_INVALID_RATE = 'clickInvalidRate';
@@ -66,6 +67,14 @@ interface StatsRepository
     ): ChartResult;
 
     public function fetchViewInvalidRate(
+        string $publisherId,
+        string $resolution,
+        DateTime $dateStart,
+        DateTime $dateEnd,
+        ?string $siteId = null
+    ): ChartResult;
+
+    public function fetchViewUnique(
         string $publisherId,
         string $resolution,
         DateTime $dateStart,