+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - User impersonation by server admin
 - Recognize fake events in statistics
-- Columns to statistics report: all views/clicks, invalid views/clicks rate
+- Columns to statistics report: all views/clicks, invalid views/clicks rate, unique views
 ### Changed
 - Formulas for computing CPC, CPM, RPC, RPM. Total cost is used instead of payment for event of particular type
 ### Fixed